[#177558088] refactoring and fixes
